refactor: moves ProviderRawData and its unit tests under a single folder
